api/requirements.txt
aiohttp==3.11.11
pydantic==2.10.6
python-dotenv==1.0.1
